+++ NEWS.md
@@ -13,7 +13,7 @@
 - replace mutate and slice(n()) with summarise in the f_dose_observed and f_dose_draw functions to improve efficiency
 - plot gap time t0 instead of t0 + 1 in the f_fit_t0 function
 - plot the rounded value of di based on probability calculations alongside the observed value of di in the f_fit_di function
-
+- replace the residual plot with the fitted gap time bar chart in the f_fit_ti function
 
 # drugDemand 0.1.0